requirements.txt
asgiref==3.3.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.18.26
botocore==1.21.26
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.45.0
django-etc==1.3.0
django-hitcount==1.3.3
django-paypal==1.1.1
django-storages==1.11.1
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
matplotlib-inline==0.1.2
moviepy==1.0.3
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.1.1
paramiko==2.7.2
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
proglog==0.1.9
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
storage==0.0.4.3
tqdm==4.62.2
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
